refactor: split quantized_vectors.rs into a module
`quantized_vectors.rs` had grown to ~2.6k lines. Split it into a
`quantized_vectors/` module:

- `config.rs`: path constants, `QuantizedVectorsConfig`,
  `QuantizedVectorsStorageType`
- `storage.rs`: `QuantizedVectorStorage` enum, type aliases and its
  `is_on_disk`/`heap_size_bytes`/`Debug` impls
- `accessors.rs`: per-variant dispatch accessors (`files`, `populate`,
  `clear_cache`, `flusher`, `upsert_vector`, ...)
- `create.rs` / `load.rs`: cross-type creation/loading dispatch
- `scalar/`, `pq/`, `binary/`, `turbo/`: per quantization type, each
  with its own `create.rs` and `load.rs`

The module root keeps the `QuantizedVectors` struct, scorer methods,
path/conversion helpers and the `MemoryReporter` impl, and re-exports
the public types so the external API path is unchanged. Code was moved
verbatim; no behavior change.
