+++ lib/segment/src/vector_storage/quantized/quantized_vectors/scalar/mod.rs
@@ -1,0 +1,2 @@
+mod create;
+mod load;
